Reorganize Content Collection config
