+++ src/config/content-collections/index.ts
@@ -1,0 +1,2 @@
+export * from './doc';
+export * from './site-config';
